integration: extract simple_strategy_test
This test is about load balancing more than about the Session.
