+++ scylla/tests/integration/load_balancing/mod.rs
@@ -1,4 +1,5 @@
 mod latency_awareness;
 mod lwt_optimisation;
 mod shards;
+mod simple_strategy;
 mod tablets;
